+++ test/unit/fuzz-tools.test.js
@@ -339,6 +339,7 @@ import * as pressureInjuryStageV359 from '../../lib/pressure-injury-stage-v359.j
 import * as kwbRetinopathyV360 from '../../lib/kwb-retinopathy-v360.js';
 import * as tannerStagingV361 from '../../lib/tanner-staging-v361.js';
 import * as forresterHemodynamicV362 from '../../lib/forrester-hemodynamic-v362.js';
+import * as shafferAngleV363 from '../../lib/shaffer-angle-v363.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -623,6 +624,7 @@ const MODULES = {
   'kwb-retinopathy-v360.js': kwbRetinopathyV360,
   'tanner-staging-v361.js': tannerStagingV361,
   'forrester-hemodynamic-v362.js': forresterHemodynamicV362,
+  'shaffer-angle-v363.js': shafferAngleV363,
   'tb-testing.js': tbTesting,
 };